feat: expose createLauncherBuilder in LSPClientFeatures for custom MessageProducer support
Extract the anonymous Launcher.Builder from LanguageServerDefinition
into DefaultLauncherBuilder with a protected createMessageProducer()
method that subclasses can override to provide a custom MessageProducer
(e.g., for non-stdio message sources).

Add LSPClientFeatures.createLauncherBuilder() so plugins can customize
the launcher without requiring a custom LanguageServerDefinition or
programmatic registration.

Changes:
- New DefaultLauncherBuilder class with protected createMessageProducer()
- LSPClientFeatures.createLauncherBuilder() returns DefaultLauncherBuilder
  by default, overridable by plugins
- LanguageServerWrapper calls through LSPClientFeatures instead of
  LanguageServerDefinition directly
- LanguageServerDefinition.createLauncherBuilder() simplified to delegate
  to DefaultLauncherBuilder
