+++ src/main/java/com/redhat/devtools/lsp4ij/server/definition/LanguageServerDefinition.java
@@ -30,19 +30,9 @@ import com.redhat.devtools.lsp4ij.client.features.LSPClientFeatures;
 import com.redhat.devtools.lsp4ij.features.semanticTokens.DefaultSemanticTokensColorsProvider;
 import com.redhat.devtools.lsp4ij.features.semanticTokens.SemanticTokensColorsProvider;
 import com.redhat.devtools.lsp4ij.installation.ServerInstaller;
-import com.redhat.devtools.lsp4ij.internal.ExtendedConcurrentMessageProcessor;
-import com.redhat.devtools.lsp4ij.internal.ExtendedStreamMessageProducer;
-import com.redhat.devtools.lsp4ij.internal.capabilities.CodeLensOptionsAdapter;
+import com.redhat.devtools.lsp4ij.server.DefaultLauncherBuilder;
 import com.redhat.devtools.lsp4ij.settings.contributors.LanguageServerSettingsContributor;
-import org.eclipse.lsp4j.CodeLensOptions;
-import org.eclipse.lsp4j.jsonrpc.*;
-import org.eclipse.lsp4j.jsonrpc.json.ConcurrentMessageProcessor;
-import org.eclipse.lsp4j.jsonrpc.json.MessageJsonHandler;
-import org.eclipse.lsp4j.jsonrpc.json.StreamMessageConsumer;
-import org.eclipse.lsp4j.jsonrpc.json.StreamMessageProducer;
-import org.eclipse.lsp4j.jsonrpc.messages.RequestMessage;
-import org.eclipse.lsp4j.jsonrpc.messages.ResponseError;
-import org.eclipse.lsp4j.jsonrpc.services.ServiceEndpoints;
+import org.eclipse.lsp4j.jsonrpc.Launcher;
 import org.eclipse.lsp4j.services.LanguageServer;
 import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.NotNull;
@@ -54,10 +44,6 @@ import java.util.Locale;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.CopyOnWriteArrayList;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-import java.util.concurrent.atomic.AtomicInteger;
-import java.util.function.Function;
 
 /**
  * Base class for Language server definition.
@@ -216,93 +202,19 @@ public abstract class LanguageServerDefinition implements LanguageServerFactory,
     }
 
     /**
-     * Custom RemoteEndpoint that uses integer IDs instead of string IDs for JSON-RPC messages.
+     * Creates the {@link Launcher.Builder} used to build the LSP launcher.
+     *
+     * <p>The default implementation returns a {@link DefaultLauncherBuilder}. Subclasses
+     * can override this to return a custom builder, but the preferred extension point
+     * is {@link LSPClientFeatures#createLauncherBuilder()} which avoids requiring
+     * a custom {@link LanguageServerDefinition}.</p>
+     *
+     * @param clientFeatures the client features.
+     * @param <S>            the language server interface type.
+     * @return the launcher builder.
      */
-    private static class RemoteEndpointWithIdAsInt extends RemoteEndpoint {
-
-        private final AtomicInteger nextRequestId = new AtomicInteger();
-
-        public RemoteEndpointWithIdAsInt(MessageConsumer out,
-                                         Endpoint localEndpoint,
-                                         Function<Throwable, ResponseError> exceptionHandler) {
-            super(out, localEndpoint, exceptionHandler);
-        }
-
-        public RemoteEndpointWithIdAsInt(MessageConsumer out,
-                                         Endpoint localEndpoint) {
-            super(out, localEndpoint);
-        }
-
-        @Override
-        protected RequestMessage createRequestMessage(String method, Object parameter) {
-            RequestMessage requestMessage = new RequestMessage();
-            // Use int as JSON-RPC id instead of String (by default from LSP4J)
-            requestMessage.setId(nextRequestId.incrementAndGet());
-            requestMessage.setMethod(method);
-            requestMessage.setParams(parameter);
-            return requestMessage;
-        }
-    }
-
     public <S extends LanguageServer> Launcher.Builder<S> createLauncherBuilder(@NotNull LSPClientFeatures clientFeatures) {
-
-        return new Launcher.Builder<S>() {
-
-            public Launcher<S> create() {
-                // Validate input
-                if (input == null)
-                    throw new IllegalStateException("Input stream must be configured.");
-                if (output == null)
-                    throw new IllegalStateException("Output stream must be configured.");
-                if (localServices == null)
-                    throw new IllegalStateException("Local service must be configured.");
-                if (remoteInterfaces == null)
-                    throw new IllegalStateException("Remote interface must be configured.");
-
-                // Create the JSON handler, remote endpoint and remote proxy
-                MessageJsonHandler jsonHandler = createJsonHandler();
-                RemoteEndpoint remoteEndpoint = createRemoteEndpoint(jsonHandler);
-                S remoteProxy = createProxy(remoteEndpoint);
-
-                // Create the message processor
-                StreamMessageProducer reader = new ExtendedStreamMessageProducer(input, jsonHandler, remoteEndpoint);
-                MessageConsumer messageConsumer = wrapMessageConsumer(remoteEndpoint);
-                ConcurrentMessageProcessor msgProcessor = createMessageProcessor(reader, messageConsumer, remoteProxy);
-                ExecutorService execService = executorService != null ? executorService : Executors.newCachedThreadPool();
-                return createLauncher(execService, remoteProxy, remoteEndpoint, msgProcessor);
-            }
-
-            @Override
-            protected ConcurrentMessageProcessor createMessageProcessor(MessageProducer reader, MessageConsumer messageConsumer, S remoteProxy) {
-                return new ExtendedConcurrentMessageProcessor(reader, messageConsumer);
-            }
-
-            @Override
-            protected RemoteEndpoint createRemoteEndpoint(MessageJsonHandler jsonHandler) {
-
-                boolean useIntAsId = clientFeatures.isUseIntAsJsonRpcId();
-                if (!useIntAsId) {
-                    // Use JSON-RPC as String (default behavior of LSP4J)
-                    return super.createRemoteEndpoint(jsonHandler);
-                }
-
-                // Override the remote endpoint to use JSON-RPC id as int
-                MessageConsumer outgoingMessageStream = new StreamMessageConsumer(output, jsonHandler);
-                outgoingMessageStream = wrapMessageConsumer(outgoingMessageStream);
-                Endpoint localEndpoint = ServiceEndpoints.toEndpoint(localServices);
-                RemoteEndpoint remoteEndpoint;
-                if (exceptionHandler == null)
-                    remoteEndpoint = new RemoteEndpointWithIdAsInt(outgoingMessageStream, localEndpoint);
-                else
-                    remoteEndpoint = new RemoteEndpointWithIdAsInt(outgoingMessageStream, localEndpoint, exceptionHandler);
-                jsonHandler.setMethodProvider(remoteEndpoint);
-                return remoteEndpoint;
-            }
-        }.configureGson(builder -> {
-            // Add a custom CodeLensOptionsAdapter to support old language server
-            // which declares codeLenProvider with a boolean instead of Json object.
-            builder.registerTypeAdapter(CodeLensOptions.class, new CodeLensOptionsAdapter());
-        });
+        return new DefaultLauncherBuilder<>(clientFeatures);
     }
 
     public boolean supportsCurrentEditMode(@NotNull Project project) {
